// src/main/resources/angular/src/app/api/model/user.ts
/**
 * MAAS API
 * This is the Mobility as a Service API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: avalls@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AppConfig } from './appConfig';
import { CreditCard } from './creditCard';
import { Permissions } from './permissions';


/**
 * All data of user
 */
export interface User {
    userId: number;
    email: string;
    locale: string;
    /**
     * Active ticket
     */
    ticketID?: number;
    documentTypeId: string;
    documentNumber: string;
    mobilePrefix: string;
    mobilePhone: string;
    fullName: string;
    accountId: number;
    appConfig?: AppConfig;
    permissions: Permissions;
    creditcard?: CreditCard;
    userStateId?: number;
    applicationType?:number;
}