// src/main/resources/angular/src/app/api/model/sproUserStatus.ts
/**
 * Smou API
 * This is the Mobility as a Service API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: avalls@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { UserStatus } from './userStatus';

/**
 * All data of user
 */
export interface SproUserStatus { 
    userId:number;
    status: UserStatus;
    unpaid? :boolean;
    unpaidsTiquets:Array<number>;
}